backend/scraper/pyproject.toml
[project]
name = "scraper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.11.11",
    "email-validator>=2.2.0",
    "fastapi>=0.115.6",
    "google-api-python-client>=2.159.0",
    "google-auth>=2.37.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.1",
    "google-generativeai>=0.8.3",
    "httpx>=0.28.1",
    "litestar[standard]>=2.14.0",
    "psycopg[binary]>=3.2.4",
    "pydantic-settings>=2.7.1",
    "python-jose>=3.3.0",
    "redis>=5.2.1",
    "schedule>=1.2.2",
    "sqlalchemy>=2.0.37",
    "watchdog>=6.0.0",
]

[dependency-groups]
dev = [
    "ruff>=0.9.2",
]
